moved useToggleDarkMode hook out of EmberBridge
